server/requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
Django==5.1.6
django-admin-interface==0.29.4
django-colorfield==0.12.0
django-cors-headers==4.7.0
django-filter==25.1
django-grappelli==4.0.1
django-jet==1.0.8
django-material==1.12.0
django-rest-framework==0.1.0
django-unfold==0.49.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pillow==11.1.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-slugify==8.0.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
yarl==1.18.3
